requirements.txt
discord.py==1.7.3
youtube_dl==2021.6.6
python-dotenv==0.19.0
youtube-search-python==1.4.9
PyNaCl==1.4.0